requirements.txt
fastapi
uvicorn
sqlalchemy
razorpay
python-dotenv
python-multipart
requests

# For Railway Postgres



